+++ src/Utils/Database.js
@@ -19,11 +19,10 @@ class Database {
     // Starts database connection
     if (database.connection) {
       if (!database.connection.useNewUrlParser) database.connection.useNewUrlParser = true
-      if (!database.connection.autoReconnect) database.connection.autoReconnect = true
       if (!database.connection.keepAlive) database.connection.keepAlive = true
       if (!database.connection.keepAliveInitialDelay) database.connection.keepAliveInitialDelay = 300000
     } else {
-      database.connection = { useNewUrlParser: true, autoReconnect: true, keepAlive: true, keepAliveInitialDelay: 300000, connectTimeoutMS: 300000 }
+      database.connection = { useNewUrlParser: true, keepAlive: true, keepAliveInitialDelay: 300000, connectTimeoutMS: 300000 }
     }
     this.#dbConnection.url = database.url
     this.#dbConnection.options = database.connection