+++ src/Utils/Server.js
@@ -9,7 +9,7 @@ const morgan = require('morgan')
 const bearerToken = require('express-bearer-token')
 
 class Server {
-  constructor (https, ssl, ENCRYPTIONKEY, logger, corsOpt) {
+  constructor (https, ssl, ENCRYPTIONKEY, logger, corsOpt, serverAddon) {
     this.app = express()
 
     this.server = false
@@ -50,6 +50,9 @@ class Server {
       reqKey: 'token',
       cookie: false
     }))
+
+    // Executing server addon
+    if (serverAddon) serverAddon(this.app)
   }
 
   listen (conf, message) {